djangoapp/requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.4
beautifulsoup4==4.6.3
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
cssselect==1.0.3
decorator==4.3.0
dj-database-url==0.5.0
Django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.8.2
fabric==2.4.0
feedfinder2==0.0.4
feedparser==6.0.10
gunicorn==19.9.0
idna==2.7
invoke==1.2.0
ipdb==0.13
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
jieba==0.39
jieba3k==0.35.1
lxml==4.2.5
newspaper3k==0.2.8
nltk==3.3
numpy==1.19
paramiko==2.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.6
psycopg2-binary==2.9
ptyprocess==0.6.0
PuLP==1.6.8
pyasn1==0.4.4
pycorenlp==0.3.0
pycparser==2.19
Pygments==2.2.0
PyNaCl==1.3.0
pyparsing==2.2.2
#git+git://github.com/banrahan/PyTeaser
python-dateutil==2.7.3
python-igraph==0.10.4
pytz==2018.5
PyYAML==3.13
requests==2.28.1
requests-file==1.5.1
scikit-learn==1.0.0
scipy==1.7.0
simplegeneric==0.8.1
six==1.11.0
tiktoken
tinysegmenter==0.3
tldextract==2.2.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
whitenoise==4.1
PyPDF2
